% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmod.R
\name{biglist}
\alias{biglist}
\title{Fuse all results from a list into one huge data frame}
\usage{
biglist(x)
}
\arguments{
\item{x}{a list (of lists (of lists (of ...)))}
}
\value{
a data frame with additional columns L0, L1, ...
}
\description{
Fuse all results from a list into one huge data frame
}
\details{
This function expect a list of an arbitrary depth, and a data frame at
the end of each branch.
}
