% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_browser.R
\name{gene_browser}
\alias{gene_browser}
\title{Launch a browser of DE analysis results}
\usage{
gene_browser(pip, cntr = NULL, annot = NULL)
}
\arguments{
\item{pip}{pipeline object returned by \code{load_de_pipeline}}

\item{cntr}{(optional) pre-loaded contrasts (returned by \code{get_contrasts})}

\item{annot}{(optional) pre-loaded annotation table (returned by \code{get_annot})}
}
\value{
does not return a value
}
\description{
Launch a shiny-based browser of DE analysis results
}
\details{
Launches a shiny app, web based, which allows to show gene expression
profiles in a pipeline.

To speed up launching, you can pre-load the contrasts with the
\code{get_contrasts} function.
}
\examples{
\dontrun{
pip <- load_de_pipeline(config_file="DE_config.yaml")
gene_browser(pip, tmod_dbs)
}
}
