% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_browser.R
\name{geneBrowserPlotUI}
\alias{geneBrowserPlotUI}
\alias{geneBrowserPlotServer}
\title{Shiny Module – gene browser expression profile plot}
\usage{
geneBrowserPlotUI(id, covar)

geneBrowserPlotServer(id, gene_id, covar, exprs, annot)
}
\arguments{
\item{id}{identifier (same as the one passed to geneBrowserTableUI)}

\item{covar}{data frame with all covariates}

\item{gene_id}{primary identifier of the gene to show}

\item{exprs}{expression matrix; row names must correspond to the primary identifiers}

\item{annot}{annotation data frame containing column 'PrimaryID'
corresponding to the rownames of the contrast data frames}
}
\value{
does not return anything useful
}
\description{
Shiny Module – gene browser expression profile plot
}
