% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_browser.R
\name{geneBrowserTableServer}
\alias{geneBrowserTableServer}
\title{Shiny Module – gene browser table selection}
\usage{
geneBrowserTableServer(id, cntr, annot)
}
\arguments{
\item{id}{identifier (same as the one passed to geneBrowserTableUI)}

\item{cntr}{a list of data frames containing the DE analysis results}

\item{annot}{annotation data frame containing column 'PrimaryID'
corresponding to the rownames of the contrast data frames}
}
\value{
reactive value returning gene ID
}
\description{
Shiny Module – gene browser table selection
}
