% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_browser.R
\name{geneBrowserTableUI}
\alias{geneBrowserTableUI}
\title{Create a shiny UI for a gene expression table browser}
\usage{
geneBrowserTableUI(id, cntr_titles)
}
\arguments{
\item{id}{identifier for the namespace of the module}

\item{cntr_titles}{named character vector for contrast choices}
}
\description{
Create a shiny UI for a gene expression table browser
}
