% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_all}
\alias{get_all}
\title{Get all objects from the pipeline}
\usage{
get_all(
  x,
  file_tab = TRUE,
  covar = TRUE,
  counts = TRUE,
  config = TRUE,
  contrasts = TRUE,
  exprs = TRUE,
  annot = TRUE,
  pca = TRUE,
  deseq2 = TRUE,
  prcomp = TRUE,
  extensions = TRUE
)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
a list with all the objects
}
\description{
Get all objects from the pipeline. This function is useful to
get all the objects from the pipeline in one go.
}
