% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_annot}
\alias{get_annot}
\title{Get the pipeline annotation}
\usage{
get_annot(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
data frame with the annotation
}
\description{
Get the annotation data frame of the pipeline
}
