% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_config}
\alias{get_config}
\title{Get config}
\usage{
get_config(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
a list object.
}
\description{
Get the configuration of pipeline read from yaml file produced in the pipeline_report step.
}
