% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_contrast_names}
\alias{get_contrast_names}
\title{Get all contrast names}
\usage{
get_contrast_names(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
A character vector with contrast names
}
\description{
Get the names of all defined contrasts from a pipeline
}
