% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_contrasts}
\alias{get_contrasts}
\title{Get all results of DE analysis}
\usage{
get_contrasts(x, contrasts = NULL)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{contrasts}{an optional character vector of the contrasts to return}
}
\value{
a named list; each object of the list is a data frame with the results of DE analysis of a given contrast.
}
\description{
Get results of the DE analysis for all contrasts (default) or a
selection of the contrasts.
}
