% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_counts}
\alias{get_counts}
\title{Get all counts from the pipeline}
\usage{
get_counts(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
a matrix with raw counts, rownames corresponding to PrimaryIDs
and column names corresponding to sample names
}
\description{
Get counts from the import_gene_counts step
}
