% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_covariates}
\alias{get_covariates}
\title{Get the covariates}
\usage{
get_covariates(x, hide = 75)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{hide}{hide all columns wider than this}
}
\value{
a color data frame with the covariates
}
\description{
Get the covariates from the DESeq2 object
}
\details{
Note: rather than reading the covariate file (which might have changed
since DESeq2 was run) the function reads the DESeq2 object and gets the
column data of that object.

Note 2: The returned object is a color data frame. Some columns are
hidden on purpose because they clutter the view. Use \code{hide=Inf} to
prevent this behavior.

Note 3: All factors are converted back to strings.
}
