% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_deseq2}
\alias{get_deseq2}
\title{Get the DESeq2 object}
\usage{
get_deseq2(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
an object of class DESeq2
}
\description{
Get the DESeq2 object created by the DE pipeline
}
