% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_exprs}
\alias{get_exprs}
\title{Get the expression data object}
\usage{
get_exprs(x, model = FALSE)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{model}{if TRUE, model rlog values are returned instead of blind rlog
values.}
}
\value{
an object of class DESeq2
}
\description{
Get the expression data object created by the DE pipeline
}
