% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_extensions}
\alias{get_extensions}
\title{Show all the extensions of the pipeline}
\usage{
get_extensions(x, steps = NULL)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{steps}{a character vector showing the steps for which the
extensions are to be shown}
}
\value{
a character vector with the extensions
}
\description{
Show the extensions (file types) the pipeline has
}
