% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_file_tab}
\alias{get_file_tab}
\title{Get the file tab}
\usage{
get_file_tab(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
A color data frame with files
}
\description{
Get a tabular representation of all output files created by the pipeline
}
\seealso{
load_de_pipeline
}
