% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_object}
\alias{get_object}
\title{Get a pipeline object}
\usage{
get_object(
  x,
  step,
  extension = "rds",
  contrast = NULL,
  multiple_ok = TRUE,
  as_list = FALSE
)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{step}{of the pipeline}

\item{extension}{of the object}

\item{contrast}{optional: choice of the contrast}

\item{multiple_ok}{Whether it is OK to return multiple objects}

\item{as_list}{Always return a list of objects}
}
\value{
an object return by readRDS, or a list of objects
}
\description{
Get a pipeline object from a given step and extension
}
