% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_object_path}
\alias{get_object_path}
\title{Get file path of a pipeline object}
\usage{
get_object_path(x, step, extension, contrast = NULL, multiple_ok = TRUE)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{step}{of the pipeline}

\item{extension}{of the object}

\item{contrast}{optional: choice of the contrast}

\item{multiple_ok}{whether it is OK to return multiple objects}
}
\value{
a character vector of file paths
}
\description{
Get file path of a pipeline object
}
