% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_pca}
\alias{get_pca}
\title{Principal component analysis}
\usage{
get_pca(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
an object of class prcomp
}
\description{
Principal component analysis
}
