% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_steps}
\alias{get_steps}
\title{Show all the steps of the pipeline}
\usage{
get_steps(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
a character vector with the steps
}
\description{
Show the steps the pipeline has
}
