% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_tmod_dbs}
\alias{get_tmod_dbs}
\title{Return the tmod gene set database object}
\usage{
get_tmod_dbs(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
a list containing gene set databases and corresponding tmod objects
}
\description{
Return the tmod gene set database object
}
