% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_tmod_gene_lists}
\alias{get_tmod_gene_lists}
\title{Return the tmod gene list object}
\usage{
get_tmod_gene_lists(x, contrasts = NULL)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{contrasts}{an optional character vector of the contrasts to return}
}
\value{
tmod gene list object
}
\description{
Return the tmod gene list object
}
\details{
The gene list object holds the gene lists used in the tmod analysis. The
object is a named list corresponding to each contrast. Each element of the
list is a named list corresponding to each tmod database. Each element of
this list is a named character vector. Names of this vector are PrimaryIDs,
and values are the corresponding IDs in the given database.
}
\seealso{
\code{\link[=get_tmod_dbs]{get_tmod_dbs()}}, \code{\link[=get_tmod_mapping]{get_tmod_mapping()}}
}
