% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_tmod_mapping}
\alias{get_tmod_mapping}
\title{Return the tmod database mapping object}
\usage{
get_tmod_mapping(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
a list containing the tmod mapping object
}
\description{
Return the tmod database mapping object
}
\details{
It is best to use the output of this function with \code{\link[=tmod_db_map_ids]{tmod_db_map_ids()}}.

tmod mapping object is a set of mappings between the identifiers used in
the tmod databases such as MSigDB and the primary identifiers used in the
pipeline. The object is a list consisting of three elements:
\itemize{
\item \code{dbs}: a named character vector. Names are the IDs of the tmod
databases; values are the IDs of the mappings.
\item \code{maps}: a named list of named character vectors. Names of the
elements of the list are mapping IDs and correspond to the values of the
\code{dbs} element. Elements of the list are mappings. Each mapping is a
named character vector. Names of this vector are PrimaryIDs, and values are
the corresponding IDs in the given database.
Return the tmod database mapping object
}
}
\seealso{
\code{\link[=get_tmod_dbs]{get_tmod_dbs()}}, \code{\link[=tmod_db_map_ids]{tmod_db_map_ids()}}
}
