% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_tmod_pca_prcomp}
\alias{get_tmod_pca_prcomp}
\title{Get the prcomp object produced by tmod_pca}
\usage{
get_tmod_pca_prcomp(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
data frame with principal components
}
\description{
Get the prcomp object produced by tmod_pca
}
