% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_utils.R
\name{get_tmod_pca_res}
\alias{get_tmod_pca_res}
\title{Get the results of tmod_pca}
\usage{
get_tmod_pca_res(x)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}
}
\value{
data frame with principal components
}
\description{
Get the results of tmod_pca
}
