% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tmod_panelplot.R
\name{gg_panelplot}
\alias{gg_panelplot}
\title{Create a tmod panel plot using ggplot}
\usage{
gg_panelplot(
  res,
  pie,
  auc_thr = 0.65,
  q_thr = 0.05,
  filter_row_pval = 0.01,
  filter_row_auc = 0.65,
  pval_cutoff = 1e-12
)
}
\description{
Create a tmod panel plot using ggplot
}
