% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_pipeline.R, R/mapping_pipeline.R
\name{print.seasnap_DE_pipeline}
\alias{print.seasnap_DE_pipeline}
\alias{load_de_pipeline}
\alias{summary.seasnap_DE_pipeline}
\alias{summary.seasnap_mapping_pipeline}
\title{Load a sea-snap DE pipeline from a directory}
\usage{
\method{print}{seasnap_DE_pipeline}(x, ...)

load_de_pipeline(config_file = "DE_config.yaml", config = NULL)

\method{summary}{seasnap_DE_pipeline}(object, ...)

\method{summary}{seasnap_mapping_pipeline}(object, ...)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{...}{Further arguments passed to the \code{print()} function}

\item{config_file}{path to the DE yaml configuration file}

\item{config}{object accessible in R snippets of the DE
pipeline containing the pipeline configuration and produced files. If tis
parameter is defined, it overrides the \code{path} parameter.}

\item{object}{an object of class seasnap_DE_pipeline}
}
\value{
seasnap DE pipeline object of class "seasnap_DE_pipeline"
}
\description{
Load a sea-snap DE pipeline from a directory
}
\details{
This is the entry point for using a sea-snap DE pipeline in R. First
create an object using this function, next use all the goodies that come
with Rseasnap.

There are two ways of creating the object: either use an existing
configuration object
(config object in R snippets) or point to a DE pipeline yaml configuration file.
In any case, the config is only consulted to get the output file path of
the pipeline; then, the files file_tab and config are read to ensure that
the object is up to date.
}
