% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_pipeline.R
\name{print.seasnap_mapping_pipeline}
\alias{print.seasnap_mapping_pipeline}
\alias{load_mapping_pipeline}
\title{Load a sea-snap mapping pipeline from a directory}
\usage{
\method{print}{seasnap_mapping_pipeline}(x, ...)

load_mapping_pipeline(config_file = "mapping_config.yaml")
}
\arguments{
\item{x}{an object of class seasnap_mapping_pipeline}

\item{...}{Further arguments passed to the \code{print()} function}

\item{config_file}{path to the mapping yaml configuration file}
}
\value{
seasnap DE pipeline object of class "seasnap_DE_pipeline"
}
\description{
Load a sea-snap mapping pipeline from a directory
}
\details{
This is the entry point for using a sea-snap mapping pipeline in R. First
create an object using this function, next use all the goodies that come
with Rseasnap.

There is only one way of creating the object:
point to a mapping pipeline yaml configuration file.
}
