% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pip_to_seapiper}
\alias{pip_to_seapiper}
\title{Create a seapiper_ds object}
\usage{
pip_to_seapiper(x)
}
\arguments{
\item{x}{a seasnap_DE_pipeline object}
}
\value{
An object of class seapiper_ds
}
\description{
Create an object that can be used with seaPiper
}
