% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_browser.R
\name{pipeline_browser}
\alias{pipeline_browser}
\title{Sea-snap pipeline browser}
\usage{
pipeline_browser(
  pip,
  annot = NULL,
  cntr = NULL,
  tmod_res = NULL,
  tmod_dbs = NULL,
  primary_id = "PrimaryID"
)
}
\arguments{
\item{pip}{pipeline returned by \code{load_de_pipeline}}

\item{annot}{annotation returned by \code{get_annot}}

\item{cntr}{contrasts list returned by \code{get_contrasts}}

\item{tmod_res}{tmod results returned by \code{get_tmod_res}}

\item{tmod_dbs}{tmod databases returned by \code{get_tmod_dbs}}

\item{primary_id}{name of the column in the annotion data frame
which corresponds to the primary gene identifier (including the
row names of the contrasts results in the cntr object)}
}
\description{
Sea-snap pipeline browser
}
\details{
Launch a shiny app with gene and tmod results browsers for a pipeline.
Launching is much faster if the large objects (contrasts, tmod
databases) do not need to be loaded each time the pipeline browser is
started.
}
\examples{
if(interactive()) {
  example_dir <- system.file("extdata/example_pipeline", package="Rseasnap")
  conf_f      <- file.path(example_dir, "DE_config.yaml")
  pip         <- load_de_pipeline(config_file = conf_f)
  pipeline_browser(pip)
}
}
