% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_disco}
\alias{plot_disco}
\title{Create a disco plot}
\usage{
plot_disco(
  contrast1,
  contrast2,
  lower = -100,
  upper = 100,
  show_top_labels = 0,
  annot = NULL
)
}
\arguments{
\item{contrast1, contrast2}{data frames with rownames corresponding to
IDs (they don't need to be in the same order) and columns \code{log2FoldChange}
and \code{pvalue}.}

\item{lower, upper}{lower and upper boundaries for coloring of the score}

\item{show_top_labels}{sort the genes by descending absolute disco score and show top N labels}

\item{annot}{annotation object returned by \code{get_annot()} or any other data
frame with columns "PrimaryID" and "SYMBOL"}
}
\value{
a ggplot object (plot)
}
\description{
Make a concordance / discordance plots for two contrasts
}
\details{
A concordance / discordance plot is simply a log~2~ FC vs log~2~ FC
comparison between two contrasts.
Disco score is a heuristic score
reflecting the strength of similarity or dissimilarity between two
log~2~ FC values weighted by the corresponding p-values and given by the
formula log2FC.x * log2FC.y * (-log10(pval.x) - log10(pval.y)).

To manually select which labels to show, use a truncated \code{annot} object
– remove all rows you don't want to show, and set the show_top_labels
parameter to \code{Inf}.
}
