% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_evidence_pca.R, R/tmod.R
\name{plot_tmod_pca_evidence}
\alias{plot_tmod_pca_evidence}
\alias{plot_evidence}
\alias{plot.evidence_plot_data}
\title{Create tmod evidence plot}
\usage{
plot_tmod_pca_evidence(
  x,
  id,
  dbname,
  pc,
  direction = "abs",
  tmod_dbs_obj = NULL,
  annot_obj = NULL,
  ...
)

plot_evidence(
  x,
  id,
  dbname,
  contrast,
  sort = "pval",
  contrast_obj = NULL,
  gl = NULL,
  tmod_dbs_obj = NULL,
  tmod_dbs_mapping_obj = NULL,
  annot_obj = NULL,
  gg = FALSE,
  ...
)

\method{plot}{evidence_plot_data}(x, id, ...)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{id}{ID of the gene set (module)}

\item{dbname}{Name of the tmod database (see get_tmod_db_names)}

\item{pc}{the ID of the principal component to show the evidence plot for}

\item{direction}{abs, up or down. "abs" means the component weights are
sorted descendingly by absolute value; "up" means they are sorted with
the smallest at the top of the list; "down" means they are sorted with
the largest values at the top of the list.}

\item{tmod_dbs_obj}{the object returned by \code{get_object(x, "tmod_dbs")}}

\item{annot_obj}{annotation data frame returned by get_annot(x)}

\item{...}{further arguments passed to the tmod::evidencePlot() function}

\item{contrast}{either a character vector of length one or a character vector of gene IDs}

\item{sort}{which sorting type (must be present in the gene list, don't worry if you don't know what that is)}

\item{contrast_obj}{a data frame with columns named PrimaryID, log2FoldChange and padj}

\item{gl}{(optional) gene list object returned by get_tmod_gl (if contrast is a character vector of length 1)}

\item{tmod_dbs_mapping_obj}{the object returned by \code{get_object(x, "tmod_dbs", "mapping.rds")}}

\item{gg}{if TRUE, the plot will be returned as a ggplot object}
}
\value{
Returns invisibly an object of class \code{evidence_plot_data} which
can be reused to quickly repeat the plot with another gene set ID using the
\code{plot.evidence_plot_data} function.
}
\description{
Create an evidence plot (ROC curve) for the given module and database
}
\details{
Evidence plots are ROC curves which visualize the strength of a gene set
enrichment based on an ordered list of genes.
For a given contrast, given gene set ID belonging to a given database,
\code{plot_evidence} produces such a plot.

The contrast parameter can either be a single character value (name of
the contrast) or a character vector of gene IDs. In
the first case, the argument is the name of the contrast to be used.
In the second case, the vector is assumed to be a vector of PrimaryIDs
(typically ENSEMBL IDs), just like the column "PrimaryID" in the
annotation data frame (returned by get_annot).

Since loading of objects can take time, the larger objects (tmod dbs,
contrasts, precomputed gene lists) can be provided as optional named arguments.
It is usually much more efficient to get the tmod database object first
and pass it as an optional argument than to load it each time this function
is called. The mapping object (tmod_dbs_mapping_obj) is less critical,
as it is usually much smaller and faster to load.

Alternatively, if you assign the output of \code{plot_evidence()} to a
variable, this variable will store the necessary object and can be used
with the \code{plot()} function to efficiently plot the data again.
}
