% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gene}
\alias{plot_gene}
\title{Show gene expression in relation to a covariate}
\usage{
plot_gene(
  x,
  id,
  xCovar,
  rld = NULL,
  annot = NULL,
  covar = NULL,
  groupBy = NA,
  colorBy = NA,
  symbolBy = NA
)
}
\arguments{
\item{x}{pipeline object returned by \code{load_de_pipeline()}}

\item{id}{PrimaryID of the gene (usually ENSEMBL ID)}

\item{xCovar}{the x covariate – column name from the covariate table}

\item{rld}{gene expression matrix to show on the y axis; rownames must
be PrimaryIDs. If NULL, the rld object from the pipeline is used.}

\item{annot}{annotation data frame (as returned by the get_annot()
function). If empty, it will be loaded.}

\item{covar}{the covariate data frame containing the column \code{xCovar}}

\item{groupBy}{name of the covariate column by which to group and connect by lines the data points}

\item{colorBy}{name of the covariate column by which to color the data}

\item{symbolBy}{name of the covariate column by which to select point symbols}
}
\value{
a ggplot2 object
}
\description{
Show gene expression in relation to a covariate
}
