% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmod.R
\name{plot_tmod_pca}
\alias{plot_tmod_pca}
\alias{plot_tmod_panelplot}
\title{Create tmod panel plot from pipeline data}
\usage{
plot_tmod_pca(
  x,
  dbname,
  type = c("absolute", "directional"),
  tmod_pca_res = NULL,
  ...
)

plot_tmod_panelplot(
  x,
  res,
  dbname,
  sel = NULL,
  contrasts_obj = NULL,
  annot_obj = NULL,
  tmod_dbs_obj = NULL,
  sort_by = "pval",
  lfc_column = "log2FoldChange",
  pval_column = "padj",
  gs_titles = NULL,
  dotplot = FALSE,
  pval_thr = 0.05,
  lfc_thr = 1,
  ...
)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{dbname}{Name of the tmod database from which the results were generated}

\item{type}{can be "absolute" (one enrichment per component) or "directional" (two enrichments per component)}

\item{tmod_pca_res}{results of tmod PCA analysis}

\item{...}{any further arguments are passed to tmodPanelPlot}

\item{res}{results of tmod gene set enrichment analysis. A list with one
element per contrast.}

\item{sel}{a character vector selecting the gene set IDs to plot (note that
thresholds still apply)}

\item{contrasts_obj}{a list with results of DE analysis (e.g. returned by \code{get_contrasts}).
The names of the elements must be identical to the names of the \code{res} parameter.}

\item{annot_obj}{annotation object returned by \code{get_annot} (if NULL, it
will be loaded from the pipeline).}

\item{tmod_dbs_obj}{tmod database object returned by \code{get_tmod_dbs} (if NULL, it
will be loaded from the pipeline).}

\item{sort_by}{sorting key}

\item{lfc_column}{Name of the column in contrasts containing the per gene log2 fold changes.}

\item{pval_column}{Name of the column in contrasts containing the per gene p-values.}

\item{gs_titles}{named character vector. Names must correspond to gene
set IDs from the tmod db object}

\item{dotplot}{if TRUE, only simple dots will be shown}

\item{pval_thr}{p-value threshold for significant genes}

\item{lfc_thr}{log2 FC  threshold for significant genes}
}
\description{
Create tmod panel plot from pipeline data
}
\details{
This is a wrapper around two tmod functions, tmodPanelPlot and
tmodDecideTests. The panel plots show the results of gene set enrichment
analysis; however, to add color according to whether genes go up or down,
additional information based on differential expression analysis is
required. Specifically, for each gene set, the number of significantly up-
and down-regulated genes must be determined. This is the job of
tmodDecideTests. plot_tmod_panelplot makes it easier in that it takes care
of that automatically based on the contrasts defined in the pipeline.
}
