% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmod.R
\name{test_gsea_tmod}
\alias{test_gsea_tmod}
\title{Run tmod gene set enrichment}
\usage{
test_gsea_tmod(
  x,
  gl,
  dbname,
  tmod_dbs_obj = NULL,
  tmod_dbs_mapping_obj = NULL,
  func = tmodCERNOtest,
  ...
)
}
\arguments{
\item{x}{an object of class seasnap_DE_pipeline}

\item{gl}{an ordered character vector with primary IDs}

\item{dbname}{name of the gene set database to use}

\item{tmod_dbs_obj}{object return by the \code{get_tmod_dbs()} function}

\item{tmod_dbs_mapping_obj}{the object returned by \code{get_object(x, "tmod_dbs", "mapping.rds")}}

\item{func}{tmod gene set enrichment function (by default, tmodCERNOtest)}

\item{...}{further arguments passed to the tmod gene set enrichment test function}
}
\description{
Run tmod gene set enrichment
}
\details{
It is usually much more efficient to get the tmod database object first
and pass it as an optional argument than to load it each time this function
is called. The mapping object (tmod_dbs_mapping_obj) is less critical,
as it is usually much smaller and faster to load.
}
\seealso{
\code{\link[tmod:tmodUtest]{tmod::tmodCERNOtest()}}, \code{\link[=get_tmod_dbs]{get_tmod_dbs()}}, \code{\link[=get_tmod_mapping]{get_tmod_mapping()}}
}
