% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmod.R
\name{test_results_tabbed_DT}
\alias{test_results_tabbed_DT}
\title{Produce a tabbed DT results table}
\usage{
test_results_tabbed_DT(res, pval_thr = 0.05, es_thr = 0.6)
}
\arguments{
\item{res}{results of a gene set enrichment test (e.g.  \code{\link[=test_gsea_tmod]{test_gsea_tmod()}})}

\item{pval_thr}{report only results with FDR < pval_thr}

\item{es_trh}{report only results with effect size > es_thr}
}
\description{
Produce a tabbed DT results table of gene set enrichment test.
}
\seealso{
\code{\link[=test_gsea_tmod]{test_gsea_tmod()}}
}
