% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{tmod_browser}
\alias{tmod_browser}
\title{Launch a browser of tmod gene set enrichment analysis results}
\usage{
tmod_browser(pip, tmod_dbs = NULL)
}
\arguments{
\item{pip}{pipeline object returned by \code{load_de_pipeline}}

\item{tmod_dbs}{tmod db object returned by get_tmod_dbs}
}
\value{
does not return a value
}
\description{
Launch a shiny-based browser of tmod gene set enrichment analysis results
}
\examples{
\dontrun{
pip <- load_de_pipeline(config_file="DE_config.yaml")
tmod_dbs <- get_tmod_dbs(pip)
tmod_browser(pip, tmod_dbs)
}
}
