% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tmod_browser.R
\name{tmodBrowserPlotUI}
\alias{tmodBrowserPlotUI}
\alias{tmodBrowserPlotServer}
\title{Shiny Module – tmod browser evidence plots}
\usage{
tmodBrowserPlotUI(id)

tmodBrowserPlotServer(id, selmod, pip, tmod_dbs, tmod_map, cntr, annot = NULL)
}
\arguments{
\item{id}{identifier (same as the one passed to geneBrowserTableUI)}

\item{selmod}{a reactive
value (e.g. returned by tmodBrowserTableServer) and a list containing
the module id, tmod dataset id, contrast id and sorting type}

\item{pip}{pipeline object returned by \code{load_de_pipeline}}

\item{tmod_dbs}{tmod gene set databases returned by \code{get_tmod_dbs()}}

\item{tmod_map}{tmod gene set ID mapping returned by \code{get_tmod_mapping()}}

\item{cntr}{list of contrast results returned by \code{get_contrasts()}}

\item{annot}{data frame containing gene annotation}
}
\value{
returns a reactive value with a selected gene identifier
}
\description{
Shiny Module – gene browser evidence plots
}
