% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tmod_browser.R
\name{tmodBrowserTableUI}
\alias{tmodBrowserTableUI}
\alias{tmodBrowserTableServer}
\title{Shiny Module – tmod results browser table selection}
\usage{
tmodBrowserTableUI(id, cntr_titles)

tmodBrowserTableServer(id, tmod_res, multilevel = FALSE)
}
\arguments{
\item{id}{identifier for the namespace of the module}

\item{cntr_titles}{possibly named character vector with contrast names}

\item{tmod_res}{results of tmod analysis, returned by \code{get_tmod_res}}

\item{multilevel}{if TRUE, the results are grouped in data sets}
}
\value{
reactive value producing a list containing the module id, contrast id, db name and sort type.
}
\description{
Shiny Module – tmod results browser table selection
}
