% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{gene_browser}
\alias{gene_browser}
\title{Launch a browser of DE analysis results}
\usage{
gene_browser(pip)
}
\arguments{
\item{pip}{pipeline object returned by \code{load_de_pipeline}}
}
\value{
does not return a value
}
\description{
Launch a shiny-based browser of DE analysis results
}
\examples{
\dontrun{
pip <- load_de_pipeline(config_file="DE_config.yaml")
gene_browser(pip, tmod_dbs)
}
}
