% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_browser.R
\name{geneBrowserTableUI}
\alias{geneBrowserTableUI}
\alias{geneBrowserTableServer}
\title{Shiny Module – gene browser table selection}
\usage{
geneBrowserTableUI(id, cntr_titles)

geneBrowserTableServer(id, cntr, annot)
}
\arguments{
\item{id}{identifier for the namespace of the module}

\item{cntr_titles}{named character vector for contrast choices}

\item{cntr}{a list of data frames containing the DE analysis results}

\item{annot}{annotation data frame containing column 'PrimaryID'
corresponding to the rownames of the contrast data frames}
}
\value{
reactive value containing the gene ID
}
\description{
Shiny Module – gene browser table selection
}
