% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_gene_browser.R
\name{geneBrowserTableUI}
\alias{geneBrowserTableUI}
\alias{geneBrowserTableServer}
\title{Shiny Module – gene browser table selection}
\usage{
geneBrowserTableUI(id, cntr_titles)

geneBrowserTableServer(id, cntr, annot, annot_linkout = NULL)
}
\arguments{
\item{id}{identifier for the namespace of the module}

\item{cntr_titles}{named character vector for contrast choices}

\item{cntr}{a list of data frames containing the DE analysis results}

\item{annot}{annotation data frame containing column 'PrimaryID'
corresponding to the rownames of the contrast data frames}

\item{annot_linkout}{a list; see Details.}
}
\value{
reactive value containing the gene ID
}
\description{
Shiny Module – gene browser table selection
}
\details{
The linkout feature (parameter \code{annot_linkout}) allows to define how the
different columns from the annotation data frame are represented as
linkouts to external data bases.

The parameter \code{annot_linkout} is a named list. Names must correspond to
columns from the annotation data frame. The elements of the list are
character strings containing URLs with the \verb{\%s} placeholder. For
example, if the column \code{ENSEMBL} contains ENSEMBL identifiers, you can
link out by specifying\preformatted{annot_linkout=list(ENSEMBL="https://www.ensembl.org/id/\%s")
}
}
