% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_browser.R
\name{pipeline_browser}
\alias{pipeline_browser}
\title{Sea-snap pipeline browser}
\usage{
pipeline_browser(
  pip,
  annot = NULL,
  cntr = NULL,
  tmod_res = NULL,
  tmod_dbs = NULL
)
}
\arguments{
\item{pip}{pipeline returned by \code{load_de_pipeline}}

\item{annot}{annotation returned by \code{get_annot}}

\item{cntr}{contrasts list returned by \code{get_contrasts}}

\item{tmod_res}{tmod results returned by \code{get_tmod_res}}

\item{tmod_dbs}{tmod databases returned by \code{get_tmod_dbs}}
}
\description{
Sea-snap pipeline browser
}
\details{
Launch a shiny app with gene and tmod results browsers for a pipeline.
Launching is much faster if the large objects (contrasts, tmod
databases) do not need to be loaded each time the pipeline browser is
started.
}
