% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_tmod_browser.R
\name{tmodBrowserPlotUI}
\alias{tmodBrowserPlotUI}
\alias{tmodBrowserPlotServer}
\title{Shiny Module – tmod browser evidence plots}
\usage{
tmodBrowserPlotUI(id)

tmodBrowserPlotServer(id, selmod, pip, tmod_dbs, tmod_map, cntr)
}
\arguments{
\item{id}{identifier (same as the one passed to geneBrowserTableUI)}

\item{selmod}{identifier of the gene set to plot}

\item{pip}{pipeline object returned by \code{load_de_pipeline}}

\item{tmod_dbs}{tmod gene set databases returned by \code{get_tmod_dbs()}}

\item{tmod_map}{tmod gene set ID mapping returned by \code{get_tmod_mapping()}}

\item{cntr}{list of contrast results returned by \code{get_contrasts()}}
}
\value{
does not return anything useful
}
\description{
Shiny Module – gene browser evidence plots
}
